requirements.txt
pandas==1.5.3
streamlit
google-generativeai
python-dotenv